--- scripts/tradingLimits/utils/modules/index.ts
@@ -1,5 +1,0 @@
-// Re-export all functions from module files
-export * from './errorHandler'
-export * from './exchangeProcessor'
-export * from './limitProcessor'
-export * from './tableFormatter'
